Augment the JDF language to support "->" (pointers).
Now, we dont have to double define variables in the JDF (once
for the type itself and once for the ddesc). Unfortunately, there
is an annoying side effect, we cannot have our "global" inherit the
name of their corresponding field in the JDF (or there will be
a conflict with the #define we generate to provide easy access to
the user to global variables). Instead, their name have prepended _g_.