Single lib for PaRSEC and DPLASMA.
Reshape the building process so that we accumulate all objects into a
single library for each one of the big projects. We keep parsec-base for
internal purposes but the base library never gets installed. Also all
PaRSEC related data collections (matrix, vector, hash) are now directly
in the PaRSEC lib.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>