Add an option to use the memory sanitizer (ASAN) from gcc
There needed to be some refactoring in the cmakery because
1. we need to apply the sanitizer flag to all subdirs of parsec
2. we need to NOT APPLY the sanitizer to anything tool/frontend
(profiling, parsec_ptgpp, etc)
3. hence we may need to apply different options to parsec-base.a and the
parsec-base objects included from parsec.so

target_xxx_options do not propagate to object libraries that are
components of parsec.so

Because of the different build requirements, a blanket statement like
add_compile_options on the parsec directory is not a valid approach to
propagate them to all object libraries

object libraries are rendered mostly obsolete for subdir management in
cmake 3.13 thanks to the proper handling of relative paths in target_sources

Thus: we target_source from data_dist and from fortran subdirs, we do
not create object libraries anymore for these subparts. This avoids
having to do fine tracking of what flag gets applied where when doing a
target_xxx_options(parsec PRIVATE ...). It will now apply to everything,
except...

We still produce an object library for parsec-base. The benefit is that
we can reuse the TARGET_OBJECTS (compile only once) if the options on
parsec and parsec-base are the same. If not, we re-add the BASE_SOURCES
to force double compilation. This selection is more maintainable because
these options are set in the same CMakeFile.txt.

Signed-off-by: Aurelien Bouteiller <bouteill@icl.utk.edu>