Fix the inplace DPLASMA compilation.
If python3 has been detected and we need to convert our precision generator we
would have done it in place (which 2to3 refuses to do). As a result, we were
trying to convert using python3 but with the python2 scripts, which led to
errors. This fixes the source dir build for DPLASMA, but I still think a VPATH
build should be the default for all users/developers.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>